backend/requirements.txt
flask>=2.3,<3.0
flask-cors>=3.0,<4.0
tensorflow==2.20.0
numpy>=2.0,<3.0
Pillow==10.4.0
matplotlib==3.10.7
reportlab>=4.0,<5.0
weasyprint
